requirements.txt
beautifulsoup4==4.8.1
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==1.11.26
django-bootstrap4==1.0.1
django-crispy-forms==1.8.0
django-heroku==0.3.1
django-registration==2.4.1
django-tinymce==2.8.0
gunicorn==20.0.0
Pillow==6.2.1
psycopg2==2.8.4
python-decouple==3.3
pytz==2019.3
soupsieve==1.9.5
sqlparse==0.3.0
whitenoise==4.1.4

